refactor: update tests and types folder structure
Refs: PREV-129
