+++ tsconfig.json
@@ -9,13 +9,10 @@
 		"jsx": "react",
 		"moduleResolution": "node",
 		"baseUrl": "src",
-		"paths": {
-			"test-utils": ["./utils/test-utils"]
-		},
 		"types": ["@testing-library/jest-dom"],
 	},
 	"include": ["src/**/*.ts", "src/**/*.tsx"],
-	"exclude": ["./node_modules/*", "src/**/*.test.*", "src/**/test*"],
+	"exclude": ["./node_modules/*", "src/**/*.test.*", "src/tests/*"],
 	"ts-node": {
 		"transpileOnly": true
 	}
